backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6
redis>=5.0.0
boto3>=1.28.0
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0
reportlab>=4.0.0
openpyxl>=3.1.0
pillow>=10.0.0
mysql-connector-python>=8.1.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
alembic>=1.12.0
